Created Shipments Controller for returning final shipping info, which combines usps and ups data. Extracted from application controller.
