+++ app/helpers/shipments_helper.rb
@@ -1,0 +1,2 @@
+module ShipmentsHelper
+end
