+++ spec/controllers/shipments_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe ShipmentsController, type: :controller do
+
+end
